:art: extract all game logic into Game class temporarily
